py/requirements.txt
ffmpeg>=1.4
kornia>=0.6.3
matplotlib>=3.5.1
numpy>=1.22.2
opencv-python>=4.5.5
pillow>=9.0.1
sklearn>=1.0.2
torch>=1.10.0
tqdm>=4.62.3